test(ui): separate sidebar delegated activity regressions
